API/requirements.txt
# Основные зависимости для AI-Gen проекта
uvicorn==0.38.0
fastapi==0.120.1
APScheduler==3.11.0
psycopg2-binary==2.9.10

# LangChain и LangGraph для работы с ИИ-агентом
langchain>=0.1.0
langgraph>=0.0.40
langchain-gigachat>=0.0.1

# GigaChat для работы с российской языковой моделью
gigachat>=0.1.0

# HTTP-запросы
requests>=2.31.0

# Парсинг HTML
beautifulsoup4>=4.12.0

# Автоматизация браузера
selenium>=4.15.0

# Валидация данных
pydantic>=2.0.0

# Работа с переменными окружения
python-dotenv>=1.0.0

# Дополнительные зависимости для Selenium
webdriver-manager>=4.0.0

# Для работы с JSON и данными
certifi>=2023.0.0
charset-normalizer>=3.0.0
idna>=3.0.0
urllib3>=2.0.0

# Для работы с атрибутами
attrs>=23.0.0

# Для работы с HTTP/1.1
h11>=0.14.0

# Для работы с результатами
outcome>=1.2.0

# Для работы с сокетами
PySocks>=1.7.1

# Для работы с отсортированными контейнерами
sortedcontainers>=2.4.0

# Для работы с CSS селекторами
soupsieve>=2.4.0

# Для работы с трио
trio>=0.22.0
trio-websocket>=0.10.0

# Для работы с веб-сокетами
websocket-client>=1.6.0

# Для работы с протоколом веб-сокетов
wsproto>=1.2.0

# Для расширений типов
typing-extensions>=4.7.0
